+++ src/com/kaylerrenslow/armaDialogCreator/control/ControlType.java
@@ -62,7 +62,7 @@ public enum ControlType {
 	/** all control types that are supported for the application */
 	public static final ControlType[] AVAILABLE_TYPES = {
 			Static, ControlsGroup, Button, ShortcutButton, Edit,
-			XSlider, Combo
+			XSlider, Combo, ListBox
 	};
 
 	private final int typeId;