+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/treeview/dataCreator/ListboxDataCreator.java
@@ -1,0 +1,43 @@
+package com.kaylerrenslow.armaDialogCreator.gui.main.treeview.dataCreator;
+
+import com.kaylerrenslow.armaDialogCreator.arma.control.ArmaControl;
+import com.kaylerrenslow.armaDialogCreator.arma.control.impl.ListboxControl;
+import com.kaylerrenslow.armaDialogCreator.arma.util.ArmaResolution;
+import com.kaylerrenslow.armaDialogCreator.control.ControlType;
+import com.kaylerrenslow.armaDialogCreator.data.DataKeys;
+import com.kaylerrenslow.armaDialogCreator.data.Project;
+import com.kaylerrenslow.armaDialogCreator.expression.Env;
+import com.kaylerrenslow.armaDialogCreator.gui.fxcontrol.treeView.EditableTreeView;
+import com.kaylerrenslow.armaDialogCreator.gui.fxcontrol.treeView.TreeItemDataCreator;
+import com.kaylerrenslow.armaDialogCreator.gui.main.popup.newControl.NewControlDialog;
+import com.kaylerrenslow.armaDialogCreator.gui.main.treeview.ControlTreeItemEntry;
+import com.kaylerrenslow.armaDialogCreator.gui.main.treeview.TreeItemEntry;
+import com.kaylerrenslow.armaDialogCreator.main.ArmaDialogCreator;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+/**
+ @author Kayler
+ @since July 23, 2017 */
+public class ListboxDataCreator implements TreeItemDataCreator<ArmaControl, TreeItemEntry> {
+	public static final ListboxDataCreator INSTANCE = new ListboxDataCreator();
+
+	@Nullable
+	@Override
+	public TreeItemEntry createNew(@NotNull EditableTreeView<ArmaControl, TreeItemEntry> treeView) {
+		NewControlDialog dialog = new NewControlDialog(ControlType.ListBox, ArmaDialogCreator.getCanvasView().isBackgroundTreeView(treeView));
+		dialog.show();
+		if (dialog.wasCancelled()) {
+			return null;
+		}
+
+		ArmaResolution resolution = DataKeys.ARMA_RESOLUTION.get(ArmaDialogCreator.getApplicationData());
+		ArmaControl control =
+				new ListboxControl(dialog.getClassName(), resolution, getEnv(), Project.getCurrentProject());
+		return new ControlTreeItemEntry(control);
+	}
+
+	private Env getEnv() {
+		return ArmaDialogCreator.getApplicationData().getGlobalExpressionEnvironment();
+	}
+}